# Davidson & Openmp fixes