--- a/include/votca/xtp/bseoperator_btda.h
+++ b/include/votca/xtp/bseoperator_btda.h
@@ -133,18 +133,19 @@ struct generic_product_impl<
      and then sort the contributions into the resulting vector
      we do the same for B
      * **/
-    Map<const MatrixXd> m_reshaped(m.data(), m.rows() / 2, m.cols() * 2);
-    MatrixXd temp = op._A * m_reshaped;
-    Map<const MatrixXd> temp_unshaped(temp.data(), m.rows(), m.cols());
+    const Map<const MatrixXd> m_reshaped(m.data(), m.rows() / 2, m.cols() * 2);
+    {
+    const MatrixXd temp = op._A * m_reshaped;
+    const Map<const MatrixXd> temp_unshaped(temp.data(), m.rows(), m.cols());
     dst.topRows(half) = temp_unshaped.topRows(half);
     dst.bottomRows(half) = -temp_unshaped.bottomRows(half);
-    temp = op._B * m_reshaped;
-
-    // create a second map because temp may have moved so temp_reshaped may
-    // point to invalid memory
-    Map<const MatrixXd> temp_unshaped2(temp.data(), m.rows(), m.cols());
+    }
+    {
+    const MatrixXd temp = op._B * m_reshaped;
+    const Map<const MatrixXd> temp_unshaped2(temp.data(), m.rows(), m.cols());
     dst.topRows(half) += temp_unshaped2.bottomRows(half);
     dst.bottomRows(half) -= temp_unshaped2.topRows(half);
+    }
   }
 };
 }  // namespace internal