--- a/include/votca/xtp/bseoperator_btda.h
+++ b/include/votca/xtp/bseoperator_btda.h
@@ -142,9 +142,9 @@ struct generic_product_impl<
     }
     {
       const MatrixXd temp = op._B * m_reshaped;
-      const Map<const MatrixXd> temp_unshaped2(temp.data(), m.rows(), m.cols());
-      dst.topRows(half) += temp_unshaped2.bottomRows(half);
-      dst.bottomRows(half) -= temp_unshaped2.topRows(half);
+      const Map<const MatrixXd> temp_unshaped(temp.data(), m.rows(), m.cols());
+      dst.topRows(half) += temp_unshaped.bottomRows(half);
+      dst.bottomRows(half) -= temp_unshaped.topRows(half);
     }
   }
 };