This PR does 3 things:
a) openmp_cuda functions which are supposed to be called inside a region have an explicit openthread id
b) Davidson restart no longer invalidates the last iteration
c) Davidson non-symmetric now corrects for slightly imaginary eigenvalues due to numerical noise, which lead to wrong results/crashes and other stuff before. This also allows us to use Gramschmidt orthogonalisation in all cases, which should speed up non-symmetric problems quite a bit. 

This is still a draft because I want to test it on large systems. Was tested and it works. 

Benchmark:
Gpumuscet 28threads +2 GPUs MKL/O3/xhost
basis 555
auxbasis 1431
full BSE 64548x64548
old version: 495s
new version: 440s
new version+GPU: 388s
New version+ GPU + new Davidson 310s

I wanted to use C++17 features, but it seems C++17 and intel 19 compiler and libint do not interact so well. So we are stuck with C++14. @junghans @JoshuaSBrown 

Fix #685