@JensWehner 2021-02-20
@votca-bot changelog: redo openmp_cuda api, correct davidson for non symmetric problems.