@baumeier 2021-02-22
From the point of view of functionality, this is all very nice. I do think though that we need some developer documentation about what is going on here since it *is* pretty advanced for someone who is not *that* much of an expert. This does extend also to how the BSE operators are set up because it is very difficult to see, which call is for which part of the Hamiltonian.