@JensWehner 2021-02-22
Okay sure, can you tag the parts, which need more documentation? The BSE Hamiltionian?